% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfiles.R
\name{allfiles}
\alias{allfiles}
\title{Load raw list of all files available to raadtools from the data repository.}
\usage{
allfiles(...)
}
\arguments{
\item{...}{reserved for future use, currently ignored}
}
\value{
\code{data.frame} with columns \code{fullname} with all file paths
}
\description{
Raw list of all files.
}

