% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amps.R
\name{amps_d1_icefiles}
\alias{amps_d1_icefiles}
\title{AMPS files}
\usage{
amps_d1_icefiles(data.source = "", time.resolution = "12hourly", ...)
}
\arguments{
\item{data.source}{ignored, reserved for future use}

\item{time.resolution}{time resolution data to read, 6hourly only for now}

\item{...}{reserved for future use, currently ignored}
}
\description{
AMPS files
}
