% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amps.R
\name{amps_d1files}
\alias{amps_d1files}
\title{AMPS files}
\usage{
amps_d1files(data.source = "", time.resolution = "4hourly", ...)
}
\arguments{
\item{data.source}{ignored, reserved for future use}

\item{time.resolution}{time resolution data to read, 6hourly only for now}

\item{...}{reserved for future use, currently ignored}
}
\description{
AMPS files
}
\section{gdalinfo}{
NA
}

