% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmp_wind.R
\name{ccmp_files}
\alias{ccmp_files}
\title{Title}
\usage{
ccmp_files(time.resolution = "6hourly", ...)
}
\arguments{
\item{time.resolution}{time resoution data to read, daily or monthly}

\item{...}{passed in to brick, primarily for \code{filename}}
}
\value{
data frame of file paths
}
\description{
Title
}
\examples{
ccmp_files()
}
