% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_L3deriv.R
\name{chla_johnsonfiles}
\alias{chla_johnsonfiles}
\title{Derived ocean colour files}
\usage{
chla_johnsonfiles(time.resolution = c("daily"), product = "MODISA")
}
\arguments{
\item{time.resolution}{temporal resolution (base is daily)}

\item{product}{which product}
}
\value{
data frame
}
\description{
For L3 bin derived products.
}
