% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla.R
\name{chlafiles}
\alias{chlafiles}
\title{Chlorophyll-a}
\usage{
chlafiles(
  time.resolution = c("weekly", "monthly"),
  product = c("johnson", "oceancolor"),
  platform = c("MODISA", "SeaWiFS"),
  ...
)
}
\arguments{
\item{time.resolution}{weekly (8day) or monthly}

\item{product}{choice of chla product, see \code{readchla}}

\item{platform}{one of "MODISA" or "SeaWiFS"}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame
}
\description{
Chlorophyll-a for the Southern Ocean
}
\details{
This function generates a list of available chlorophyll-a files, including SeaWiFS and MODIS.
}
