% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms.R
\name{cpolarfiles}
\alias{cpolarfiles}
\title{Circumpolar ROMS files.}
\usage{
cpolarfiles(...)
}
\arguments{
\item{...}{ignored}
}
\value{
data frame of fullname, date
}
\description{
These were an experiment provided by Stuart Corney of UTAS, consider deprecated.
}
\examples{
cpolarfiles()
}
