% Generated by roxygen2 (4.0.2): do not edit by hand
\name{currentsfiles}
\alias{currentsfiles}
\title{AVISO ocean currents files}
\usage{
currentsfiles(time.resolution = c("daily"), ...)
}
\arguments{
\item{time.resolution}{time resolution to load}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of AVISO current data
}
\details{
A data.frame of file names and dates
}
\seealso{
\code{\link{readcurr}}
}

