% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivaadc.R
\name{derivaadcproducts}
\alias{derivaadcproducts}
\title{List all derived data products available through \code{\link{readderivaadc}}}
\usage{
derivaadcproducts()
}
\value{
character vector of product names
}
\description{
List all derived data products available through \code{\link{readderivaadc}}
}
\seealso{
\code{\link{readderivaadc}} and \code{\link{derivaadcfiles}}
}
