% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagemap.R
\name{imagemap}
\alias{imagemap}
\title{Maps of places}
\usage{
imagemap(map = c("ibcso_background_hq"), fact = 1L)
}
\arguments{
\item{map}{name of the map to load}

\item{fact}{resize factor, see \code{\link[terra]{aggregate}}}
}
\value{
SpatRaster, with R G B bands
}
\description{
Read map images.
}
\details{
The ancient lost art of cartography.

ibcso_background: The IBCSO RGB-map rasterlayer in high resolution
}
\references{
\url{http://www.ibcso.org/data.html}
}
