% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{data}
\name{nuyina}
\alias{nuyina}
\title{Nuyina underway track}
\format{
\code{aurora} A data frame with 9 columns.  The columns represent
longitude latitude date_time_utc air_pressure port_longwave_irradiance port_air_temperature precipitation_rate sea_water_salinity sea_water_temperature
}
\description{
Voyage track data from the Nuyina
}
\details{
This is a sample of the Nuyina underway in 2023.
}
\examples{
range(nuyina$date_time_utc)
}
\references{
see data-raw for extraction
}
\keyword{data}
