% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_L3deriv.R
\name{oc_sochla_files}
\alias{oc_sochla_files}
\title{Derived ocean colour files}
\usage{
oc_sochla_files(time.resolution = c("daily"), product = c("MODISA",
  "SeaWiFS"))
}
\arguments{
\item{time.resolution}{temporal resolution (base is daily)}

\item{product}{which product}
}
\value{
data frame
}
\description{
For L3 bin derived products.
}
