% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocfiles.R
\name{ocfiles}
\alias{ocfiles}
\title{Load metadata and location of files of ocean colour data products.}
\usage{
ocfiles(time.resolution = c("daily", "weekly", "monthly", "weekly32"),
  product = c("MODISA", "SeaWiFS", "VIIRS"), varname = c("RRS", "CHL",
  "POC", "KD490", "NPP_PAR", "SNPP_CHL", "SNPP_RRS"), type = c("L3b", "L3m"),
  bz2.rm = TRUE, ext = c("nc", "main"), ...)
}
\arguments{
\item{time.resolution}{daily or monthly files?}

\item{product}{choice of ocean colour sensor}

\item{varname}{which variable (or set of variables)}

\item{type}{which level of data}

\item{bz2.rm}{ignore files that are compressed}

\item{ext}{which extension, e.g. "nc" or "main"}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of \code{file} and \code{date}
}
\description{
This function loads the latest cache of stored files for
NASA ocean colour products.
}
