% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarmap.R
\name{polar_map}
\alias{polar_map}
\title{Polar map}
\usage{
polar_map(crs = commonprojections$polar)
}
\arguments{
\item{crs}{PROJ.4 string}
}
\value{
SpatVector of lines
}
\description{
Spatial lines in the standard (71S) South Polar projection. Data is
the 'rnaturalearth ne_coastline' at 'scale = 10'.
}
\details{
Use 'crs' to modify the projection, or set to 'NA' to leave in longlat.
}
\examples{
plot(polar_map())
}
