% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prodfiles}
\alias{prodfiles}
\title{Arrigo production files}
\usage{
prodfiles(fromCache = TRUE, ...)
}
\arguments{
\item{fromCache}{load file catalog from cache, or rebuild it}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of Arrigo production data.
}
\details{
These are 8 day estimates from MODIS and SeaWiFS satellite data,
original NASA algorithm.
}

