% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_grid.R
\name{query_grid}
\alias{query_grid}
\title{Create a query grid}
\usage{
query_grid(lon, lat, date)
}
\arguments{
\item{lon}{vector of longitudes in the grid}

\item{lat}{vector of latitudes in the grid}

\item{date}{vector of dates in the grid}
}
\description{
Create a grid of query values for of longitude, latitude, and date.
}
\details{
This be used as the second argument to  \code{\link[=extract]{extract()}} for general
'point-in-time' value extraction.

The ouput is a data frame of the grid expanded, a row for every combination of
longitude, latitude, date  - so can be quite large (!) so be careful with very
fine resolution or long series.
}
\examples{
lon <- c(100, 110, 120)
lat <-seq(-60, -40, by = 5)
dts <- seq(as.Date("2002-03-03"), by = "5 days", length.out = 10)
qd <- query_grid(lon = lon, lat = lat, date = dts)
## use bilinear to ensure we get an interpolate value in the grid
qd$sst <- extract(readsst, qd, method = "bilinear")
## ssha
qd$ssha <- extract(readssh, qd, ssha = TRUE, method = "bilinear")
qd$wind_mag <- extract(readwind, qd, magonly = TRUE, method = "bilinear") 
}
