% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid_response.R
\name{rapid_responsefiles}
\alias{rapid_responsefiles}
\title{Load metadata and location of files of Rapid Response imagery.}
\usage{
rapid_responsefiles(product = c("aqua", "terra"), ...)
}
\arguments{
\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of \code{file} and \code{date}
}
\description{
This function loads the latest cache of stored files for
rapid response products.
}
\details{
We acknowledge the use of Rapid Response imagery from the Land 
Atmosphere Near-real time Capability for EOS (LANCE) system operated by 
the NASA/GSFC/Earth Science Data and Information System (ESDIS) with funding provided by NASA/HQ.
}
\examples{
\dontrun{
rf <- rapid_responsefiles()
range(rf$date)
}
}
