% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_altimetry.R
\name{read_adt_daily}
\alias{read_adt_daily}
\alias{read_ugosa_daily}
\alias{read_ugos_daily}
\alias{read_sla_daily}
\alias{read_vgos_daily}
\alias{read_vgosa_daily}
\alias{read_err_daily}
\title{Altimetry products.}
\usage{
read_ugosa_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_ugos_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_sla_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_vgos_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_vgosa_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_err_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed in to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\value{
a raster layer
}
\description{
Functions \code{read_sla_daily} and so on for "ugosa, adt, ugos, sla, vgos, vgosa, err".
}
\details{
The labels 'u', 'v' are used by another function \code{readcurr()} which reads the 'ugos' and 'vgos' components and corresponds to an older
but analogous dataset.

'ugos' is Absolute geostrophic velocity: zonal component, 'surface_geostrophic_eastward_sea_water_velocity' in m/s.

'vgos' is Absolute geostrophic velocity: meridian component, 'surface_geostrophic_northward_sea_water_velocity' in m/s.

'ugosa' is Geostrophic velocity anomalies: zonal component, 'surface_geostrophic_eastward_sea_water_velocity_assuming_sea_level_for_geoid' in m/s.

'vgosa' is Geostrophic velocity anomalies: meridian component, 'surface_geostrophic_northward_sea_water_velocity_assuming_sea_level_for_geoid' in m/s.

'adt' is Absolute dynamic topography, 'sea_surface_height_above_geoid' in metres

'sla' is Sea level anomaly, 'sea_surface_height_above_sea_level' in metres.

For the raw files see \code{raadfiles::altimetry_daily_files} and for example do

'files <- read_sla_daily(returnfiles = TRUE); ncdf4::nc_open(files$fullname\link{1})' which will give a full NetCDF metadata print out of the source files (all the variables are in each file).
}
\examples{
a <- read_adt_daily(sort(Sys.Date() - 1:50),
xylim = extent(100, 150, -70, -40))
\dontrun{
animate(a, pause = 0,
col = colorRampPalette(c("dodgerblue", "white", "firebrick"))(21),
breaks = c(seq(min(cellStats(a, min)), 0, length = 11),
           seq(0.001, max(cellStats(a, max)), length = 10)))
}
}
