% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_amsr_ssmi.R
\name{read_amsr2_3k_ice}
\alias{read_amsr2_3k_ice}
\title{Read AMSR2 sea ice data (3km.}
\usage{
read_amsr2_3k_ice(
  date,
  xylim = NULL,
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}}

\item{setNA}{mask zero and values greater than 100 as NA}

\item{latest}{if TRUE and date input is missing, return the latest time available otherwise the earliest}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\description{
Using the TIF files (these are 2012 onwards)
}
\details{
This function relies on the file-listing of \code{\link[raadfiles:amsr2_3k_daily_files]{raadfiles::amsr2_3k_daily_files()}}
}
\seealso{
raadfiles::amsr2_3k_daily_files  read_amsr2_ice
}
