% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_amsr_ssmi.R
\name{read_amsr_ice}
\alias{read_amsr_ice}
\title{Read AMSR sea ice data.}
\usage{
read_amsr_ice(date, xylim = NULL, latest = TRUE, ...)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}}

\item{latest}{if TRUE return the latest time available, ignoring the 'date' argument}

\item{...}{passed to brick, primarily for \code{filename}}
}
\description{
Both eras are available in the same series, scaling is applied to the first
series to ensure the data is always in percentage (0, 100).
}
\seealso{
raadfiles::amsr_daily_files
}
