% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_amsr_ssmi.R
\name{read_cersat_ice}
\alias{read_cersat_ice}
\title{Read CERSAT daily sea ice data.}
\usage{
read_cersat_ice(
  date,
  xylim = NULL,
  latest = TRUE,
  setNA = TRUE,
  ...,
  returnfiles = FALSE,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}}

\item{latest}{if TRUE and date input is missing, return the latest time available otherwise the earliest}

\item{setNA}{mask zero and values greater than 100 as NA}

\item{...}{passed to brick, primarily for \code{filename}}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{inputfiles}{input the files data base to speed up initialization}
}
\description{
This is southern hemisphere daily 12.5km SSM/I since 1991.
}
\seealso{
raadfiles::cersat_daily_files read_amsr_ice
}
