% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_leads.R
\name{read_leads_clim}
\alias{read_leads_clim}
\alias{read_leads_clim_south}
\alias{read_leads_clim_north}
\title{Relative lead frequencies for the polar oceans}
\usage{
read_leads_clim_south(xylim = NULL)

read_leads_clim_north(xylim = NULL)

read_leads_clim(hemisphere = c("south", "north"), xylim = NULL, ...)
}
\description{
Average Lead-Frequency for the Arctic for winter months November-April 2002/03-2018/19 based on daily lead composites as
derived from MOD/MYD-29 IST 5 min granules
}
\examples{
read_leads_clim()
read_leads_clim_north(xylim = terra::ext(c(-1, 1, -1, 1) * 50000))
south <- read_leads_clim_south()

## hone in on Mawson
pt <- reproj::reproj_xy(cbind(62 + 52/60, -(67 +  36/60)), terra::crs(south), source = "+proj=longlat")
lead <- read_leads_clim_south(xylim = terra::ext(pt[1] + c(-1, 1) * 250000, pt[2] + c(-1, 1) * 250000))
plot(lead, col = grey.colors(100))
abline(v = pt[1], h = pt[2])
}
\references{
F. Reiser, S. Willmes, G. Heinemann (2020), A new algorithm for daily sea ice lead identification in the Arctic and Antarctic
winter from thermal-infrared satellite imagery, Data subm.
}
