% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_L3deriv.R
\name{read_oc_sochla}
\alias{read_oc_sochla}
\title{Southern Ocean Chlorophyll-a}
\usage{
read_oc_sochla(date, time.resolution = c("daily"), bins = NULL,
  product = c("MODISA", "SeaWiFS"), latest = TRUE, returnfiles = FALSE,
  ..., inputfiles = NULL)
}
\arguments{
\item{date}{date/s to read}

\item{time.resolution}{daily for now}

\item{latest}{return only the latest day}

\item{returnfiles}{give the files instead of the data}

\item{...}{ignored}

\item{inputfiles}{speed up the read by inputting the files}

\item{xylim}{extent in longitude latitude using 'raster::extent'}
}
\value{
tibble of the values in L3 bin form
}
\description{
Read Southern Ocean Chlorophyll-a  in L3 bin form.
}
\details{
relies on processing done here https://github.com/AustralianAntarcticDivision/ocean_colour
}
