% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_par.R
\name{read_par}
\alias{read_par}
\title{Read PAR from NASA ocean colour.}
\usage{
read_par(
  date,
  time.resolution = "8D",
  xylim = NULL,
  lon180 = FALSE,
  nobsonly = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  inputfiles = NULL,
  ...
)
}
\arguments{
\item{date}{date or dates of data to read}

\item{time.resolution}{'8D' currently, use the NASA tokens for time period}

\item{xylim}{spatial extents to crop from source data}

\item{lon180}{if TRUE (default), convert Pacific (0,360) to Atlantic (-180,180) view}

\item{latest}{if TRUE and date missing, return latest available}

\item{returnfiles}{if TRUE, return the file catalog instead of data}

\item{inputfiles}{optional pre-filtered file catalog}

\item{...}{passed to underlying reader}
}
\value{
raster object
}
\description{
Currently using MODISA 4km Mapped L2m, see \code{\link[raadfiles:par]{raadfiles::par_files()}} for the actual source files.
}
\examples{
read_par(latest = FALSE)
}
