% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_altimetry.R
\name{read_ugosa_daily}
\alias{read_ugosa_daily}
\title{Altimetry products.}
\usage{
read_ugosa_daily(date, xylim = NULL, latest = TRUE,
  returnfiles = FALSE, ..., inputfiles = NULL)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{latest}{if TRUE return the latest time available, ignoring the 'date' argument}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed in to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\value{
a raster layer
}
\description{
Functions `read_sla_daily` and so on for "ugosa, adt, ugos, sla, vgos, vgosa, err".
}
\details{
sla is sea level anomaly, for the raw files see `raadfiles::altimetry_daily_files`
}
\examples{
a <- read_adt_daily(sort(Sys.Date() - 1:50), 
xylim = extent(100, 150, -70, -40))
\dontrun{
animate(a, pause = 0, 
col = colorRampPalette(c("dodgerblue", "white", "firebrick"))(21), 
breaks = c(seq(min(cellStats(a, min)), 0, length = 11), 
           seq(0.001, max(cellStats(a, max)), length = 10)))
}
}
