% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe.R
\name{readcafe}
\alias{readcafe}
\title{Read 'cafe' MODIS monthly data}
\usage{
readcafe(
  date,
  time.resolution = c("monthly"),
  xylim = NULL,
  lon180 = TRUE,
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[terra]{ext}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180]}

\item{setNA}{mask out land values (only applies to monthly time.resolution)}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed to \code{\link[terra]{writeRaster}}, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\description{
I don't remember what these are but they were at https://orca.science.oregonstate.edu/
}
\details{
Files are found with \code{raadfiles::cafe_monthly_files()}.
}
\examples{
readcafe(date = "2010-01-15")
}
