% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl32rolling.r
\name{readchl32}
\alias{readchl32}
\title{Read R32 VIIRS files}
\usage{
readchl32(date, time.resolution = c("weekly32"), product = c("oceancolor"),
  platform = c("VIIRS"), xylim = NULL, returnfiles = FALSE,
  latest = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{date}{}

\item{time.resolution}{}

\item{product}{}

\item{platform}{}

\item{xylim}{}

\item{returnfiles}{}

\item{latest}{}

\item{verbose}{}

\item{...}{}
}
\description{
Read R32 VIIRS files
}

