% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_L3deriv.R
\name{readchla_johnson}
\alias{readchla_johnson}
\title{Read Johnson L3 bins}
\usage{
readchla_johnson(date, time.resolution = c("daily"), xylim = NULL,
  latest = FALSE, returnfiles = FALSE, ..., inputfiles = NULL)
}
\arguments{
\item{date}{date/s to read}

\item{time.resolution}{daily for now}

\item{xylim}{ignored}

\item{latest}{return only the latest day}

\item{returnfiles}{give the files instead of the data}

\item{...}{ignored}

\item{inputfiles}{speed up the read by inputting the files}
}
\value{
tibble of the values in L3 bin form
}
\description{
Read Johnson L3 bins
}
\details{
relies on processing done here 'file.path(getOption("default.datadir"), "data_local/acecrc.org.au/ocean_colour")'
}
\examples{
 x <- readchla_johnson(date = "2012-01-01")
 ##plot(roc::bin2lonlat(x$bin_num, 4320), col = palr::chlPal(x$chla_johnson), pch = ".")
}
