% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivaadc.R
\name{readderivaadc}
\alias{readderivaadc}
\title{Read data from polar climatological and other summary environmental data}
\usage{
readderivaadc(products, xylim = NULL, returnfiles = FALSE, ...)
}
\arguments{
\item{products}{choice of products, see \code{\link{derivaadcproducts}} for available products}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[terra]{ext}}}

\item{returnfiles}{ignore options and just return the file names}

\item{...}{passed on, primarily for \code{filename}}
}
\value{
\code{SpatRaster}
}
\description{
Derived data are read from files managed by \code{\link{derivaadcfiles}}.
}
\examples{
\dontrun{
prods <- c("bathymetry","chl_summer_climatology")
x <- readderivaadc(prods)
}
}
\references{
\url{http://data.aad.gov.au/aadc/metadata/metadata.cfm?entry_id=Polar_Environmental_Data}
}
\seealso{
\code{\link{derivaadcfiles}} for details on the repository of
data files
}
