% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fronts.R
\name{readfronts}
\alias{readfronts}
\title{Read data from the Sokolov/Rintoul Southern Ocean fronts analysis.}
\usage{
readfronts(
  date,
  time.resolution = c("weekly"),
  product = c("sokolov"),
  xylim = NULL,
  lon180 = TRUE,
  setNA = FALSE,
  trim = TRUE,
  returnfiles = FALSE,
  RAT = TRUE,
  ...
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{product}{choice of product, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[terra]{ext}}, see Details}

\item{lon180}{if TRUE, data originally in Pacific view will be returned in Atlantic view (-180:180)}

\item{setNA}{is \code{TRUE} NaN values are replaced with NA}

\item{trim}{if \code{TRUE} the map is cropped to no contiguous margins of missing data}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{RAT}{if \code{TRUE} data is returned as a categorical raster with the region names as its levels, see \code{\link[terra]{levels}}}

\item{...}{reserved for future use, currently ignored}
}
\value{
\code{SpatRaster}
}
\description{
Sokolov Rintoul
}
\examples{
\dontrun{
b <- readfronts(c("1993-01-01", "2005-01-02"), lon180 = FALSE)
}
}
