% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mld.R
\encoding{ISO-8859-2}
\name{readmld}
\alias{readmld}
\title{Mixed Layer Depth}
\usage{
readmld(date, xylim = NULL, returnfiles = FALSE, ...)
}
\arguments{
\item{date}{date to extract, can only be a climatology month}

\item{xylim}{extent specification}

\item{returnfiles}{return the file details only}

\item{...}{ignored}
}
\value{
RasterBrick
}
\description{
Read the \enc{Sall?e}{Sallee} Mixed layer depth climatology.
}
\details{
http://soki.aad.gov.au/display/Data/Mixed+layer+depth
}
