% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod.R
\name{readprod}
\alias{readprod}
\title{Arrigo production data}
\usage{
readprod(date, time.resolution = "weekly", xylim = NULL,
  returnfiles = FALSE, ...)
}
\arguments{
\item{date}{date or dates of data to read}

\item{time.resolution}{choice of temporal resolution, weekly only}

\item{xylim}{crop or not}

\item{returnfiles}{if TRUE return just the files from \code{prodfiles}}

\item{...}{reserved for future use, currently ignored}
}
\value{
RasterLayer or RasterBrick
}
\description{
Read Arrigo production data.
}
\details{
Arrigo production on Stereographic grid.
}
