% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid_response.R
\name{readrapid_response}
\alias{readrapid_response}
\title{MODIS Rapid Response images}
\usage{
readrapid_response(
  date,
  product = c("aqua", "terra"),
  latest = TRUE,
  returnfiles = FALSE,
  ...
)
}
\arguments{
\item{date}{date of image to load}

\item{latest}{if TRUE and date input is missing return the latest time available, otherwise the earliest}

\item{returnfiles}{return just the list of files}

\item{...}{\code{filename} to write the result to, otherwise ignored}
}
\value{
SpatRaster
}
\description{
Read MODIS Rapid Response RGB images
}
\details{
MODIS Rapid Response, Antarctica
}
