% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{readtopo}
\alias{readtopo}
\alias{topofile}
\alias{readbathy}
\alias{read_rema_tiles}
\title{Topography data}
\usage{
readtopo(
  topo = c("gebco_23", "gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m", "cryosat2",
    "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "rema_1km", "rema_200m",
    "rema_100m", "rema_8m", "gebco_19", "gebco_21"),
  polar = FALSE,
  lon180 = TRUE,
  xylim = NULL,
  returnfiles = FALSE,
  ...,
  resample = "bilinear"
)

readbathy(
  topo = c("gebco_23", "gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m", "cryosat2",
    "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "rema_1km", "rema_200m",
    "rema_100m", "rema_8m", "gebco_19", "gebco_21"),
  polar = FALSE,
  lon180 = TRUE,
  xylim = NULL,
  returnfiles = FALSE,
  ...,
  resample = "bilinear"
)

topofile(
  topo = c("gebco_23", "gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m", "cryosat2",
    "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "rema_1km", "rema_200m",
    "rema_100m", "rema_8m", "gebco_19", "gebco_21"),
  polar = FALSE,
  lon180 = TRUE,
  ...
)

read_rema_tiles(...)
}
\arguments{
\item{topo}{Data source, see Details.}

\item{polar}{Flag for returning the polar version of the IBCSO data.}

\item{lon180}{Flag for returning data in Atlantic \link{-180, 180} rather than Pacific \link{0, 360} view.}

\item{xylim}{spatial extents or raster grid to crop from source data, see Details}

\item{returnfiles}{Return just the relevant file name}

\item{...}{reserved for future use, ignored currently}

\item{resample}{method to use for GDAL warper resampling (currently via terra), the 'method' argument to 'project()'}
}
\value{
\describe{
\item{}{\code{topofile} returns a character string of the full path to a file name}
\item{}{\code{readtopo} and \code{readbathy} return the requested data as a RasterLayer (these are aliases)}
}
}
\description{
Functions to provide topographic (bathymetry and/or topography) data.
}
\details{
Use \code{readtopo} (or its alias \code{readbathy}) to read data
from the chosen data set. The function \code{topofile} is used to
find the full file name.

\code{readtopo()} and \code{readbathy()}  accept \code{terra::ext} or
\code{raster::extent} or just numeric objects for 'xylim'. Alternatively these can be
a terra \code{SpatRaster} or a raster \code{BasicRaster} (RasterLayer, RasterBrick, or
RasterStack) as a template target raster for crop and resize, or reprojection to
new raster grid.  The 'resample' argument controls the kind of sampling when
regridded or warped. (i.e. if xylim is extent, you get crop(), if it's a grid,
it is remodelled to the grid. extent doesn't need crs, grid does).
The following data sets are available using the argument \code{topo}.
\describe{
\item{gebco_23}{The GEBCO grid, 2023 version "ice surface"}
\item{gebco_08}{The GEBCO_08 Grid, a global 30 arc-second grid largely generated by combining quality-controlled ship depth soundings with interpolation between sounding points guided by satellite-derived gravity data. \url{http://www.gebco.net/data_and_products/gridded_bathymetry_data/}}
\item{ibcso}{IBCSO bathymetry data, resolution 1min, use argument \code{polar = TRUE} to return the projected version (polar stereographic with true scale at 71S, WGS84), 500m resolution. \url{http://www.ibcso.org/data.html}. Default is Ice Surface 'ibcso_is', use 'ibcso_bed' for Bedrock.}
\item{etopo1}{ETOPO1 is a 1 arc-minute global relief model of Earth's surface that integrates land topography and ocean bathymetry. \url{http://www.ngdc.noaa.gov/mgg/global/global.html}}
\item{etopo2}{Historic and deprecated prior version of ETOPO1. \url{http://www.ngdc.noaa.gov/mgg/global/etopo2.html}}
\item{kerguelen}{Kerguelen Plateau Bathymetric Grid,  \url{https://data.gov.au/dataset/ds-ga-a05f7893-007f-7506-e044-00144fdd4fa6/details?q=}}
\item{george_v_terre_adelie}{A bathymetric Digital Elevation Model (DEM) of the George V and Terre Adelie continental shelf and margin - 100, 250, and 500 metre resolution. \url{http://data.aad.gov.au/aadc/metadata/metadata_redirect.cfm?md=AMD/AU/GVdem_2008}}
\item{smith_sandwell}{Global seafloor topography from satellite altimetry and ship depth soundings. \url{http://topex.ucsd.edu/WWW_html/mar_topo.html}}
\item{cryosat2}{Antarctica CryoSat-2 Digital Elevation Model (DEM). \url{https://earth.esa.int/web/guest/missions/esa-operational-eo-missions/cryosat}}
\item{lake_superior}{Bathymetry of Lake Superior \url{https://www.ngdc.noaa.gov/mgg/greatlakes/superior.html}}
\item{ramp}{Radarsat Antarctic digital elevation model V2 \url{https://github.com/AustralianAntarcticDivision/blueant#radarsat-antarctic-digital-elevation-model-v2}}
\item{ga_srtm}{Digital Elevation Model (DEM) of Australia with 1 Second Grid}
\item{rema_1km, rema_200m, rema_100m, rema_8m}{Reference Elevation Model of Antartica (REMA) for the peninsula, see \code{read_rema_tiles} for the index}
\item{gebco_19}{GEBCO_2019 Grid https://www.gebco.net/data_and_products/gridded_bathymetry_data/gebco_2019/gebco_2019_info.html}
}
}
