% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity.R
\name{salfiles}
\alias{salfiles}
\title{SMAP sea surface temperature files}
\usage{
salfiles(time.resolution = c("daily"), ...)
}
\arguments{
\item{time.resolution}{time resolution read}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of SMAP sea surface salinity
}
\details{
A data frame of file names and dates.

Only a short period of daily data is available, from '2015-03-27' to
'2018-10-21'.
}
