% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssh.R
\name{sshfiles}
\alias{sshfiles}
\title{AVISO sea surface height / anomaly files}
\usage{
sshfiles(time.resolution = c("daily", "monthly", "monthly_clim",
  "seasonal_clim"), ssha = FALSE, ...)
}
\arguments{
\item{time.resolution}{time resolution to find}

\item{ssha}{logical value, return absolute (SSH) or relative (SSHA anomaly) values}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of AVISO SSH/SSHA data
}
\details{
A data.frame of file names and dates
}
\seealso{
\code{\link{readssh}}
}

