% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datapal.R
\name{sst.pal}
\alias{sst.pal}
\title{SST colours}
\usage{
sst.pal(x, palette = FALSE, alpha = 1)
}
\arguments{
\item{x}{a vector of data values or a single number}

\item{palette}{logical, if \code{TRUE} return a list with matching colours and values}

\item{alpha}{value in 0,1 to specify opacity}
}
\value{
colours, palette, or function, see Details
}
\description{
SST colours
}
\references{
Derived from \url{"http://oceancolor.gsfc.nasa.gov/DOCS/palette_sst.txt}.
}

