% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{sstfiles}
\alias{sstfiles}
\title{OISST sea surface temperature files}
\usage{
sstfiles(
  time.resolution = c("daily", "monthly"),
  varname = c("sst", "ice", "err", "anom"),
  ...
)
}
\arguments{
\item{time.resolution}{time resolution read}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of OISST sea surface temperature data
}
\details{
A data frame of file names and dates
}
