% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations locations}
\description{
In polar stereographic, with longitude and latitude and name as attributes.
}
\details{
Obtained with antanym
}
\examples{
plot(readice())
plot(stations, add = TRUE)
}
