% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tables.R
\name{table_uvgos}
\alias{table_uvgos}
\title{Read surface currents as table}
\usage{
table_uvgos(
  date,
  xylim = NULL,
  ...,
  xy = TRUE,
  cell = FALSE,
  na.rm = TRUE,
  latest = TRUE,
  res = NULL
)
}
\arguments{
\item{date}{dates to read}

\item{xylim}{extent of data to read}

\item{...}{arguments passed to read functions (only \code{lon180} relevant)}

\item{xy}{include coordinates of cell in the output, \code{TRUE} by default}

\item{cell}{include cell index in the output, \code{FALSE} by default}

\item{na.rm}{by default missing values are removed, set to \code{FALSE} to keep all}
}
\value{
data frame of u, v, x,y (longitude,latitude), cell, and date
}
\description{
Read meridional and zonal components of surface currents (m/s) from altimetry products. Input
is optional, with \code{date} and \code{xylim}. As with the grid reading functions \link{read_ugos_daily} and \link{read_vgos_daily}
this by default will return the latest data available and for the entire world.
}
\details{
Please note that the coordinates are in longitude latitude, but the velocity components are in m/s. You cannot
meaningfully transform the x,y coordinates and use the velocity components without taking into account
rotation in the transformation (we might write some helpers for this ...).

Argument \code{lon180} may be used to specify Pacific or Atlantic orientation.
}
\examples{
uv <- table_uvgos("2001-01-01", xylim = terra::ext(60, 120, -60, -10))
plot(range(uv$x), range(uv$y), type = "n", asp = 1.1)
scal <- function(x) (x - min(x, na.rm = TRUE))/diff(range(x, na.rm = TRUE))
nn <- 56
arrows(uv$x, uv$y, uv$x + uv$u, uv$y + uv$v, col = grey.colors(nn)[scal(sqrt(uv$u^2 + uv$v^2)) * (nn-1) + 1], length = 0)
}
