% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raadtools.R
\name{timedateFrom}
\alias{timedateFrom}
\title{Stable conversion to POSIXct from character and Date}
\usage{
timedateFrom(x, ...)
}
\arguments{
\item{x}{input date-time stamp, character, Date or other supported type.}

\item{\dots}{ignored}
}
\value{
the vector \code{x} converted (if necessary) to \code{POSIXct}
}
\description{
Conversion to POSIXct ensuring no local time zone applied. Currently supported is character, Date and
anything understood by \code{\link[base]{as.POSIXct}}.
}

