% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{topofile}
\alias{topofile}
\title{Find topography data files}
\usage{
topofile(
  topo = c("gebco_23", "gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m", "cryosat2",
    "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "rema_1km", "rema_200m",
    "rema_100m", "rema_8m", "gebco_19", "gebco_21"),
  polar = FALSE,
  lon180 = TRUE,
  ...
)
}
\arguments{
\item{topo}{Data source, see \code{\link{read_topo}}.}

\item{polar}{Flag for returning the polar version of the IBCSO data.}

\item{lon180}{Flag for returning data in Atlantic \link{-180, 180} rather than Pacific \link{0, 360} view.}

\item{...}{reserved for future use, ignored currently}
}
\value{
character string of the full path to a file name
}
\description{
Return the full path, or GDAL data source name, for a topography or
bathymetry data set. Use \code{\link{read_topo}} to read the data itself.
}
\details{
The available data sets, and what each one is, are listed in the Details
section of \code{\link{read_topo}}.
}
\seealso{
\code{\link{read_topo}} to read the data
}
