% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{windfiles}
\alias{windfiles}
\title{Files containing NCEP2 wind vector data}
\usage{
windfiles(data.source = "", time.resolution = c("6hourly"), ...)
}
\arguments{
\item{data.source}{ignored, reserved for future use}

\item{time.resolution}{time resolution data to read, 6hourly only for now}

\item{...}{reserved for future use, currently ignored}
}
\value{
\code{data.frame} of file names and dates
}
\description{
NCEP2 wind files
}
\details{
Files containing NCEP2 wind vector data
}
