% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{data}
\name{amps_metadata}
\alias{amps_metadata}
\title{AMPS GRIB file metadata}
\format{
\code{amps_metadata} A data frame with 8 columns.  The columns
represent
\tabular{rl}{
\code{Band} \tab the band number \cr
\code{GRIB_COMMENT} \tab the data measured \cr
\code{GRIB_ELEMENT} \tab the data name \cr
\code{GRIB_FORECAST_SECONDS} \tab forecast seconds  \cr
\code{GRIB_REF_TIME} \tab text reference time \cr
\code{GRIB_SHORT_NAME} \tab short name  \cr
\code{GRIB_UNIT} \tab  text unit \cr
\code{GRIB_VALID_TIME} \tab text valid time \cr
}
}
\description{
GRIB format metadata from the Antarctic Mesoscale Prediction System (AMPS) files.
}
\examples{
print(amps_metadata)
## u_wind_at_900 <- read  ## unfinished
}
\keyword{data}
