% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chlafiles}
\alias{chlafiles}
\title{Chlorophyll-a}
\usage{
chlafiles(time.resolution = c("weekly", "monthly"), product = c("johnson",
  "oceancolor"), ...)
}
\arguments{
\item{time.resolution}{weekly (8day) or monthly}

\item{product}{choice of sea ice product, see \code{readchla}}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame
}
\description{
Chlorophyll-a for the Southern Ocean
}
\details{
This function generates a list of available chlorophyll-a files, including SeaWiFS and MODIS.
}

