% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{data}
\name{commonprojections}
\alias{commonprojections}
\title{This is a list of often used projections, in PROJ}
\format{
An object of class \code{list} of length 4.
}
\usage{
commonprojections
}
\description{
Each element can be looked up by name, see Examples
}
\section{Warning}{

This should be use only for a convenient reference to look up the projection strings commonly in use. There's
no guarantee that this would be appropriate and you should seek cartographic expertise.
}

\examples{
names(commonprojections)
commonprojections[["polar"]]
}
\references{
\url{http://www.spatialreference.org}
}
\seealso{
\code{\link[raster]{projection}}, \code{\link[sp]{CRS}}, \code{\link[sp]{proj4string}}
}
\keyword{data}
