% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivaadc.R
\name{derivaadcfiles}
\alias{derivaadcfiles}
\title{Load metadata and location of files of polar climatological and other summary environmental data}
\usage{
derivaadcfiles(products, ...)
}
\arguments{
\item{products}{which derived product}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of \code{file} and \code{date}
}
\description{
This function loads the latest cache of stored files for
these products, which are available from \url{http://webdav.data.aad.gov.au/data/environmental/derived/antarctic/}
}

