% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deriv.R
\name{derivicefiles}
\alias{derivicefiles}
\title{Load metadata and location of files of derived sea ice data products.}
\usage{
derivicefiles(product = "time_since_melt", ...)
}
\arguments{
\item{product}{which derived product}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of \code{file} and \code{date}
}
\description{
This function loads the latest cache of stored files for
ice products, currently only daily NSIDC southern hemisphere is available.
}

