% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_ice.R
\name{distance_to_ice_edge}
\alias{distance_to_ice_edge}
\alias{distance_to_ice}
\title{Distance to a sea ice 'edge'.}
\usage{
distance_to_ice_edge(
  date,
  threshold = 15,
  xylim = NULL,
  hemisphere = "south",
  returnfiles = FALSE,
  inputfiles = NULL,
  latest = TRUE,
  ...
)

distance_to_ice(
  date,
  threshold = 15,
  xylim = NULL,
  hemisphere = "south",
  returnfiles = FALSE,
  inputfiles = NULL,
  latest = TRUE,
  ...
)
}
\arguments{
\item{date}{date or dates of data to read}

\item{threshold}{the sea ice concentration threshold to contour at}

\item{xylim}{spatial extents to crop, or a raster template for warping}

\item{hemisphere}{"north" or "south", default is "south"}

\item{returnfiles}{if TRUE, return file catalog}

\item{inputfiles}{optional pre-filtered file catalog}

\item{latest}{if TRUE and date missing, return latest available}

\item{...}{passed to underlying reader}
}
\value{
raster layer with distances to this date's sea ice edge
}
\description{
Calculate the shortest distance (metres) to a threshold sea ice contour.  If in
doubt use \code{distance_to_ice_edge}, the definition of the edge is not straightforward, especially so for
the higher resolution products and near the coast.
\code{distance_to_ice_edge} computes a single "main" edge at continental scale
\code{distance_to_ice} computes all distances to any ice at threshold concentration
}
\details{
The distance is always positive, use \code{readice} in the usual way to determine if a
location is inside or out of the ice field itself. (If inside means zero distance to ice
for you then set it explicitly based on the concentration a point is in.)
Future work may generalize this to other data sources.
}
\note{
beware that any queried location outside of this layer's range will be
undetermined, and the external boundary of this layer is not constant with
respect to the pole, and that in general a location may be closer to ice in the
opposite hemisphere.

The argument \code{hemisphere} may be north or south (default is south), but this will only work if your locations
are on the actual map, so it's not possible to request the distance to ice in both poles for any point.
}
\examples{
plot(distance_to_ice(latest = TRUE))
plot(distance_to_ice_edge(latest = TRUE))
a = extract(distance_to_ice, aurora[17:25, ])
extract(distance_to_ice, aurora[17:25, ], hemisphere = "south")
# library(trip)
# extract(distance_to_ice_edge, walrus818[seq(50, 400, by = 20), ], hemisphere = "north")
}
