% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastice.R
\name{fasticefiles}
\alias{fasticefiles}
\title{Fast ice files}
\usage{
fasticefiles(...)
}
\arguments{
\item{...}{ignored}
}
\value{
RasterLayer
}
\description{
Provided by Alex Fraser
}
\details{
This function wraps a raadfiles function \code{fasticefiles}, but expands the file list out to individual bands.
}
\examples{
fraser_fasticefiles()
}
