% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastice.R
\name{fasticefiles}
\alias{fasticefiles}
\title{fast ice files}
\usage{
fasticefiles(product = "binary_fast_ice", ...)
}
\arguments{
\item{...}{reserved for future use, currently ignored}

\item{datadir}{data repository path}
}
\value{
data frame
}
\description{
Data frame of all available fast ice files.
}
\details{
A data frame with file, date, fullname
}

