% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fasticefiles}
\alias{fasticefiles}
\title{fast ice files}
\usage{
fasticefiles(datadir = getOption("default.datadir"), ...)
}
\arguments{
\item{datadir}{data repository path}

\item{...}{reserved for future use, currently ignored}
}
\value{
data frame
}
\description{
Data frame of all available fast ice files.
}
\details{
A data frame with file, date, fullname
}

