% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastice.R
\name{fasticefiles}
\alias{fasticefiles}
\title{fast ice files}
\usage{
fasticefiles(product = "binary_fast_ice", mask = FALSE, ...)
}
\arguments{
\item{product}{which product}

\item{mask}{if TRUE return mask file name}

\item{...}{reserved for future use, currently ignored}
}
\value{
data frame
}
\description{
Data frame of all available fast ice files.
}
\details{
A data frame with file, date, fullname
}
