% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsi.R
\name{frontsmap}
\alias{frontsmap}
\title{Fronts map data for the Southern Ocean}
\usage{
frontsmap(map = c("orsi"))
}
\arguments{
\item{map}{name of map to load}
}
\value{
SpatialLinesDataFrame
}
\description{
Load spatial map of fronts data.
}
\details{
Currently ORSI is the only supported layer.

"orsi" - the ORSI fronts derived from the files provided by the
WOCE Atlas, see References
}
\references{
\url{http://woceatlas.tamu.edu/Sites/html/atlas/SOA_DATABASE_DOWNLOAD.html}
}
