% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{icefiles}
\alias{icefiles}
\title{Load metadata and location of files of sea ice data products.}
\usage{
icefiles(time.resolution = "daily", product = "nsidc",
  hemisphere = c("south", "north"), ...)
}
\arguments{
\item{time.resolution}{daily or monthly files?}

\item{product}{choice of sea ice product, see \code{\link{readice}}}

\item{hemisphere}{north or south}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of \code{file} and \code{date}
}
\description{
This function loads the latest cache of stored files for
ice products.
}
\examples{
\dontrun{
icf <- icefiles(time.resolution = "monthly")
icf[which.min((as.Date("1995-01-01") + runif(1, -4000, 4000)) - as.Date(icf$date), ]
}
}
