% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{package}
\name{raadtools-package}
\alias{raadtools-package}
\title{R tools for spatial data, extensions using terra to read and extract}
\description{
Tools in R for reading, plotting and manipulating spatial data, originally
used at the Australian Antarctic Division (AAD).
}
\details{
read functions like \code{\link{readsst}} will read a data set by date-time vector, with a
set of shared arguments that work the same and documented against this dummy function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/raadtools}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/raadtools/issues}
}

}
\author{
Michael D. Sumner \email{michael.sumner@aad.gov.au}

Options

There is an option 'raadtools.check.file.exists' that is TRUE by default, set to FALSE might make very long time series requests a  bit faster.

\code{options(raadtools.check.file.exists = FALSE)}

Maintainer: Michael D. Sumner \email{michael.sumner@aad.gov.au}
}
\keyword{package}
