% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{raadtools}
\alias{raadtools}
\alias{raadtools-package}
\title{R tools for spatial data, extensions using raster to read and extract}
\description{
Tools in R for reading, plotting and manipulating spatial data, originally
used at the Australian Antarctic Division (AAD).
}
\author{
Michael D. Sumner \email{mdsumner@gmail.com}

Maintainer: Michael D. Sumner \email{mdsumner@gmail.com}
}
\keyword{package}

