% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadtools.R
\docType{package}
\name{raadtools}
\alias{raadtools}
\alias{raadtools-package}
\title{R tools for spatial data, extensions using raster to read and extract}
\description{
Tools in R for reading, plotting and manipulating spatial data, originally 
used at the Australian Antarctic Division (AAD).
}
\author{
Michael D. Sumner \email{michael.sumner@aad.gov.au}

Maintainer: Michael D. Sumner \email{michael.sumner@aad.gov.au}
}
\keyword{package}
