% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_amsr_ssmi.R
\name{read_amsr2_ice}
\alias{read_amsr2_ice}
\title{Read AMSR2 sea ice data (6km)}
\usage{
read_amsr2_ice(
  date,
  xylim = NULL,
  latest = TRUE,
  ...,
  setNA = TRUE,
  returnfiles = FALSE,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read}

\item{xylim}{spatial extents to crop, or a SpatRaster (or Raster*) template
to warp onto when \code{hemisphere = "both"}}

\item{latest}{if TRUE and date missing, return latest available}

\item{...}{passed to underlying reader}

\item{setNA}{mask special values as NA}

\item{returnfiles}{if TRUE, return file catalog}

\item{inputfiles}{optional pre-filtered file catalog}
}
\value{
\code{SpatRaster}
}
\description{
Using the TIF files (these are 2012 onwards)
}
\details{
This function relies on the file-listing of \code{\link[raadfiles:amsr2_daily_files]{raadfiles::amsr2_daily_files()}}.
Concentration is returned as a fraction (0-1).
}
\seealso{
raadfiles::amsr2_daily_files
}
