% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmp_wind.R, R/compat-ccmp.R
\name{read_ccmp}
\alias{read_ccmp}
\title{Read CCMP wind files, from the RSS Cross-Calibrated Multi-Platform Ocean Surface Wind Project}
\usage{
read_ccmp(
  date,
  time.resolution = c("6hourly"),
  xylim = NULL,
  lon180 = FALSE,
  magonly = FALSE,
  dironly = FALSE,
  uonly = FALSE,
  vonly = FALSE,
  nobsonly = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_ccmp(
  date,
  time.resolution = c("6hourly"),
  xylim = NULL,
  lon180 = FALSE,
  magonly = FALSE,
  dironly = FALSE,
  uonly = FALSE,
  vonly = FALSE,
  nobsonly = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read}

\item{time.resolution}{time resolution, currently only "6hourly"}

\item{xylim}{spatial extents to crop}

\item{lon180}{if TRUE, use Atlantic-centered longitude (default FALSE)}

\item{magonly}{return wind speed only}

\item{dironly}{return wind direction only}

\item{uonly}{return U component only}

\item{vonly}{return V component only}

\item{nobsonly}{return number of observations only}

\item{latest}{if TRUE and date missing, return latest}

\item{returnfiles}{if TRUE, return file catalog}

\item{...}{passed to underlying reader}

\item{inputfiles}{optional pre-filtered file catalog}
}
\value{
\code{RasterStack} or \code{RasterLayer}, or tibble if \code{returnfiles = TRUE}
}
\description{
RSS CCMP_RT V2.1 derived surface winds (Level 3.0)

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{read_ccmp} is superseded by \code{\link{read_ccmp_wind_6hourly}},
which returns terra \code{SpatRaster} objects.
}
\details{
CCMP wind data is read from files managed by
\code{\link{ccmp_files}}. Dates are matched to file names by
finding the nearest match in time within a short duration. By
default only one time step is returned with both U and V
components. Multiple dates can be returned for magnitude or
direction, U or V only or N-obs only.

This is the " RSS VAM 6-hour analyses starting from the NCEP GFS wind analyses"

See References.
}
\examples{
## read a single time slice, direction only
x <- read_ccmp(dironly = TRUE)


## get a local extent for a zoom plot and plot the directions [0,360) as an image with arrows
e <- extent(projectExtent(raster(extent(130, 150, -50, -30), crs = "+proj=longlat"), projection(x)))
x <- crop(read_ccmp(), e)
crds <- coordinates(x)
scale <- 0.05
vlen <- function(x) sqrt(x[[1]]^2 + x[[2]]^2)
plot(vlen(crop(x, e)))
x1 <- crds[,1]
y1 <- crds[,2]
x2 <- crds[,1] + values(x[[1]]) * scale
y2 <- crds[,2] + values(x[[2]]) * scale
arrows(x1, y1, x2, y2, length = 0.02)

## faster if we get the file list first
ccfiles <- ccmp_files()
earliest <- read_ccmp(ccfiles$date[1:16], xylim = e, magonly = TRUE, inputfiles = ccfiles)
plot(earliest, col = hcl.colors(64), zlim = c(0, 20))
}
\references{
\url{http://www.remss.com}

\url{http://www.remss.com}
}
\seealso{
\code{\link{icefiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value

\code{\link{read_ccmp_wind_6hourly}} for modern terra-based reader
}
