% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla.R
\name{read_chla_daily}
\alias{read_chla_daily}
\alias{read_chla_weekly}
\alias{read_chla_monthly}
\title{Read daily chlorphyll-a}
\usage{
read_chla_daily(
  date,
  time.resolution = "daily",
  xylim = NULL,
  lon180 = TRUE,
  varname = c("chlor_a"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_chla_weekly(
  date,
  xylim = NULL,
  lon180 = TRUE,
  varname = c("chlor_a"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_chla_monthly(
  date,
  xylim = NULL,
  lon180 = TRUE,
  varname = c("chlor_a"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read}

\item{time.resolution}{time resolution: "daily" or "monthly"}

\item{xylim}{spatial extents to crop from source data}

\item{lon180}{if TRUE (default), convert Pacific (0,360) to Atlantic (-180,180) view}

\item{varname}{variable to return from the data files, default is
"sst" or "anom", "err", "ice"}

\item{setNA}{mask out land values (ignored in new implementation, kept for API compatibility}

\item{latest}{if TRUE and date missing, return latest available}

\item{returnfiles}{if TRUE, return the file catalog instead of data}

\item{...}{passed to underlying reader}

\item{inputfiles}{optional pre-filtered file catalog}
}
\value{
raster object
}
\description{
Data is read as daily files in a way consistent with other read functions that can
be used by \code{extract()}. (\code{readchla} for example cannot be used this way as it returns a mean value for the period asked for).
}
\details{
This function will read from the entire era available to SeaWiFS, MODISA, and VIIRS.

To read from particular eras use the output of \code{ocfiles()} for inputfiles  argument.
}
\examples{
read_chla_daily(latest = FALSE) ## we should see SeaWiFS
read_chla_daily() ## we should see MODISA (or VIIRS)
}
