% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla.R
\name{read_chla_daily}
\alias{read_chla_daily}
\title{Read daily chlorphyll-a}
\usage{
read_chla_daily(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  lon180 = TRUE,
  varname = c("chlor_a"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view \link{0, 360} data to Atlantic view \link{-180, 180}}

\item{varname}{variable to return from the data files, default is
"sst" or "anom", "err", "ice"}

\item{setNA}{mask out land values (only applies to monthly time.resolution)}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed in to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\description{
Data is read as daily files in a way consistent with other read functions that can
be used by \code{extract()}. (\code{readchla} for example cannot be used this way as it returns a mean value for the period asked for).
}
\details{
This function will read from the entire era available to SeaWiFS, MODISA, and VIIRS.

To read from particular eras use the output of \code{ocfiles()} for inputfiles  argument.
}
\examples{
read_chla_daily(latest = FALSE) ## we should see SeaWiFS
read_chla_daily() ## we should see MODISA (or VIIRS)
}
