% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoid.R
\name{read_geoid}
\alias{read_geoid}
\title{Read Earth Gravitation Model}
\usage{
read_geoid(xylim = NULL, force = FALSE)
}
\arguments{
\item{xylim}{an extent, in longlat, or an object that provides one}

\item{force}{if TRUE ignore cached virtual saved file, and recompute (try this if it otherwise fails)}
}
\value{
RasterLayer, longitude latitue grid of geoid level
}
\description{
Global 2.5 Minute Geoid Undulations, a big raster from multiple files. We just create a temporary
virtual raster from the source files, and return as a RasterLayer.
}
\details{
Each file is an ESRI GRID raster data set of 2.5-minute geoid undulation values covering a 45 x 45 degree area.
Each raster file has a 2.5-minute cell size and is a subset of the global 2.5 x 2.5-minute grid of pre-computed
geoid undulation point values found on the EGM2008-WGS 84 Version web page. This ESRI GRID format represents a
continuous surface of geoid undulation values where each 2.5-minute raster cell derives its value from the original
pre-computed geoid undulation point value located at the SW corner of each cell.
}
\examples{
\dontrun{
geoid <- read_geoid()
}
}
