% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_par.R
\name{read_par}
\alias{read_par}
\title{Read PAR from NASA ocean colour.}
\usage{
read_par(
  date,
  time.resolution = "8D",
  xylim = NULL,
  lon180 = FALSE,
  nobsonly = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  inputfiles = NULL,
  ...
)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{time.resolution}{'8D' currently, use the NASA tokens for time period}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view \link{0, 360} data to Atlantic view \link{-180, 180}}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{inputfiles}{input the files data base to speed up initialization}

\item{...}{passed in to brick, primarily for \code{filename}}
}
\value{
raster object
}
\description{
Currently using MODISA 4km Mapped L2m, see \code{\link[raadfiles:read_par]{raadfiles::read_par()}} for the actual source files.
}
\examples{
read_par(latest = FALSE)
}
