% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{read_rema_tiles}
\alias{read_rema_tiles}
\title{REMA tile index}
\usage{
read_rema_tiles(...)
}
\arguments{
\item{...}{reserved for future use, ignored currently}
}
\value{
SpatialPolygonsDataFrame of REMA tile outlines
}
\description{
Read the index of Reference Elevation Model of Antarctica (REMA) tiles, as
polygon outlines. Use \code{\link{read_topo}} with one of the 'rema_'
data sets to read the elevation data itself.
}
\seealso{
\code{\link{read_topo}}
}
