% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sose.R
\name{read_sose}
\alias{read_sose}
\title{Read SOSE Southern Ocean State Estimate}
\usage{
read_sose(
  date,
  time.resolution = c("monthly"),
  varname = "",
  level = 1L,
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{varname}{variable to return from the data files, default is whatever first value
from [sose_monthly_varnames()]}

\item{level, }{defaults to 1L - there are 52}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
Model data read from files managed by
\code{\link[raadfiles]{sose_monthly_files}}. Dates are matched to file names by finding
the nearest match in time within a short duration. If \code{date}
is greater than length 1 then the sorted set of unique matches is
returned.
}
\details{
This function doesn't support xylim, lon180, or more than one date being read.
}
\examples{
\dontrun{
read_sose(varname = "Uvel", level = 1)
read_sose(varname = "SeaIceArea", level = 1, latest = FALSE)
read_sose(varname = "Chl", level = 10, latest = FALSE)

}
}
