% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srtm.R
\name{read_srtm_tile}
\alias{read_srtm_tile}
\title{Read a single SRTM tile}
\usage{
read_srtm_tile(tile_x = 39, tile_y = 6)
}
\description{
Read a tile by x and y index.
}
\details{
See \code{\link[raadfiles:srtm]{raadfiles::srtm_files()}} for filename in \code{fullname} and \code{x}, \code{y} for tile indices.
}
\keyword{internal}
