% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-topo.R
\name{read_topo}
\alias{read_topo}
\alias{read_bathy}
\alias{readtopo}
\alias{readbathy}
\title{Read topography or bathymetry data}
\usage{
read_topo(
  topo = c("gebco_23", "gebco_21", "gebco_19", "gebco_14", "gebco_08", "etopo1",
    "etopo2", "ibcso", "ibcso_is", "ibcso_bed", "smith_sandwell", "kerguelen",
    "cryosat2", "ramp", "rema_8m", "rema_100m", "rema_200m", "rema_1km", "lake_superior",
    "macrie1100m", "macrie2100m"),
  xylim = NULL,
  lon180 = TRUE,
  polar = FALSE,
  resample = "bilinear",
  returnfiles = FALSE,
  ...
)

read_bathy(
  topo = c("gebco_23", "gebco_21", "gebco_19", "gebco_14", "gebco_08", "etopo1",
    "etopo2", "ibcso", "ibcso_is", "ibcso_bed", "smith_sandwell", "kerguelen",
    "cryosat2", "ramp", "rema_8m", "rema_100m", "rema_200m", "rema_1km", "lake_superior",
    "macrie1100m", "macrie2100m"),
  xylim = NULL,
  lon180 = TRUE,
  polar = FALSE,
  resample = "bilinear",
  returnfiles = FALSE,
  ...
)

readtopo(
  topo = c("gebco_23", "gebco_21", "gebco_19", "gebco_14", "gebco_08", "etopo1",
    "etopo2", "ibcso", "ibcso_is", "ibcso_bed", "smith_sandwell", "kerguelen",
    "cryosat2", "ramp", "rema_8m", "rema_100m", "rema_200m", "rema_1km", "lake_superior",
    "macrie1100m", "macrie2100m"),
  xylim = NULL,
  lon180 = TRUE,
  polar = FALSE,
  resample = "bilinear",
  returnfiles = FALSE,
  ...
)

readbathy(
  topo = c("gebco_23", "gebco_21", "gebco_19", "gebco_14", "gebco_08", "etopo1",
    "etopo2", "ibcso", "ibcso_is", "ibcso_bed", "smith_sandwell", "kerguelen",
    "cryosat2", "ramp", "rema_8m", "rema_100m", "rema_200m", "rema_1km", "lake_superior",
    "macrie1100m", "macrie2100m"),
  xylim = NULL,
  lon180 = TRUE,
  polar = FALSE,
  resample = "bilinear",
  returnfiles = FALSE,
  ...
)
}
\arguments{
\item{topo}{character, dataset name (see Details)}

\item{xylim}{extent to crop, or SpatRaster template for warping.
Accepts anything \code{terra::ext()} understands: numeric vector
\code{c(xmin, xmax, ymin, ymax)}, SpatExtent, or SpatRaster.
If a SpatRaster with CRS is provided, data is warped to match.}

\item{lon180}{logical, return Atlantic-centered \link{-180,180} view? Default TRUE.}

\item{polar}{logical, for IBCSO return polar stereographic version?}

\item{resample}{character, resampling method for warping (default "bilinear")}

\item{returnfiles}{logical, return file path instead of data?}

\item{...}{currently ignored}
}
\value{
\code{SpatRaster}, or character file path if \code{returnfiles = TRUE}
}
\description{
Read data from various topography/bathymetry sources.
}
\details{
The following datasets are available via the \code{topo} argument:

\describe{
\item{gebco_23}{GEBCO 2023 grid (ice surface), 15 arc-second global}
\item{gebco_21}{GEBCO 2021 grid}
\item{gebco_19}{GEBCO 2019 grid}
\item{gebco_14}{GEBCO 2014 grid}
\item{gebco_08}{GEBCO 2008 grid, 30 arc-second}
\item{etopo1}{ETOPO1 1 arc-minute global relief}
\item{etopo2}{ETOPO2 (historic, deprecated)}
\item{ibcso, ibcso_is}{IBCSO v2 ice surface, 500m Antarctic}
\item{ibcso_bed}{IBCSO v2 bedrock}
\item{smith_sandwell}{Global seafloor from satellite altimetry}
\item{kerguelen}{Kerguelen Plateau bathymetric grid}
\item{cryosat2}{CryoSat-2 Antarctic DEM}
\item{ramp}{RADARSAT Antarctic DEM v2}
\item{rema_8m, rema_100m, rema_200m, rema_1km}{REMA Antarctic DEMs}
\item{lake_superior}{Lake Superior bathymetry}
\item{macrie1100m, macrie2100m}{Macquarie Ridge 100m grids}
}
}
\examples{
\dontrun{
# Read GEBCO 2023
bathy <- read_topo("gebco_23")

# Crop to region
bathy <- read_topo("gebco_23", xylim = c(100, 150, -60, -40))

# Warp to template raster
template <- terra::rast(nrows = 100, ncols = 100,
                        xmin = 100, xmax = 150, ymin = -60, ymax = -40,
                        crs = "EPSG:4326")
bathy <- read_topo("gebco_23", xylim = template)
}
}
\seealso{
\code{\link{topofile}} for file paths only
}
