% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla.R
\name{readchla}
\alias{readchla}
\alias{readCHL_month}
\title{Read Chlorophyll-a, NASA algorithm}
\usage{
readCHL_month(date, xylim = NULL, ..., inputfiles = NULL, latest = TRUE)

readchla(
  date,
  product = c("MODISA", "SeaWiFS", "VIIRS"),
  xylim = NULL,
  algorithm = c("nasa"),
  latest = TRUE,
  grid = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, ignored if grid is provided
@param algorithm nasa only}

\item{latest}{if TRUE (and date not supplied) return the latest time available, otherwise the earliest
Note that reaCHL_month reads the NASA algorith L3m products.}

\item{product}{choice of product, see Details}

\item{grid}{template raster object for output}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
Ocean colour Chlorophyll-a data, provide an input of daily dates and these will be averaged into one layer.
}
\examples{
readCHL_month()
\dontrun{
d <- readchla(c("2003-01-01", c("2003-06-01")),
         xylim = extent(100, 150, -70, -30))
}
}
\seealso{
readCHL_month

\code{\link{chlafiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}
