% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla.R
\name{readchla}
\alias{readchla}
\title{Read Chlorophyll-a for the Southern Ocean}
\usage{
readchla(date, product = c("MODISA", "SeaWiFS"), xylim = NULL,
  algorithm = c("johnson", "nasa"))
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{product}{choice of product, see Details}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}
@param algorithm johnson or nasa}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
Ocean colour Chlorophyll-a data. Default is to read from the Johnson Improved
chlorophyll-a estimates using Southern Ocean-specific calibration
algorithms, but the original MODIS and SeaWIFs products are also available via the argument \code{product}.
}
\details{
Dates are matched to file names by finding the nearest match in
time within a short duration. If \code{date} is greater than
length 1 then the sorted set of unique matches is returned.
}
\examples{
\dontrun{
d <- readchla(c("2003-01-01", c("2003-06-01")),
         xylim = extent(100, 150, -70, -30))
}
}
\references{
Johnson, R, PG Strutton, SW Wright, A McMinn, and KM
Meiners (2013) Three improved satellite chlorophyll algorithms for
the Southern Ocean, J. Geophys. Res. Oceans, 118,
doi:10.1002/jgrc.20270
\url{http://onlinelibrary.wiley.com/doi/10.1002/jgrc.20270/full}
}
\seealso{
\code{\link{chlafiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}
