% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/currents.R
\name{readcurr}
\alias{readcurr}
\title{Read AVISO ocean current data}
\usage{
readcurr(date, time.resolution = c("daily"), xylim = NULL, lon180 = TRUE,
  magonly = FALSE, dironly = FALSE, uonly = FALSE, vonly = FALSE,
  latest = FALSE, returnfiles = FALSE, ...)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resolution to read}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180]}

\item{magonly}{return just the magnitude from the U and V
components}

\item{dironly}{return just the direction from the U and V, in degrees N=0, E=90, S=180, W=270}

\item{uonly}{return just the U component of velocity}

\item{vonly}{return just the V component of velocity
components, in degrees (0 north, 90 east, 180 south, 270 west)}

\item{latest}{if TRUE return the latest time available, ignoring the 'date' argument}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed to brick, primarily for \code{filename}}
}
\value{
\code{\link[raster]{raster}} object with the "U"
(horizontal/X) and "V" (vertical/Y) components of velocity in
cm/s. Setting either of the (mutually exclusive) \code{magonly}
and \code{dironly} arguments returns the magnitude (in cm) or
direction (in degrees relative to North) of the velocity vectors.
}
\description{
Current data is read from files managed by
\code{\link{currentsfiles}}. Dates are matched to file names by
finding the nearest match in time within a short duration. By
default only one time step is returned with both U and V
components. Multiple dates can be returned for magnitude or
direction, U or V only.
}
\details{
This is the "SSALTO/DUACS - DT Geostrophic Velocities - Up-to-date Global Processing". See References.
}
\note{
These data are stored in longitude/latitude projection on the sphere between longitudes in the Pacific
view \[0, 360\], the default behaviour is to reset this to Atlantic
view \[-180, 180\] with \code{lon180}.
}
\examples{
\dontrun{
## read a single time slice, and plot the directions [0,360) as an image with arrows
x <- readcurr(dironly = TRUE)
## get a local extent for a zoom plot
e <- extent(projectExtent(raster(extent(130, 150, -50, -30), crs = "+proj=longlat"), projection(x)))
x1 <- crop(readcurr(), e)
crds <- coordinates(x1)
scale <- 2000
plot(crop(x, e))
x1 <- crds[,1]
y1 <- crds[,2]
x2 <- crds[,1] + values(x1[["U"]]) * scale
y2 <- crds[,2] + values(x1[["V"]]) * scale
arrows(x1, y1, x2, y2, length = 0.03)
}
}
\references{
\url{http://www.aviso.oceanobs.com/en/data/products/sea-surface-height-products/global/index.html}
}
\seealso{
\code{\link{icefiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}

