% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivaadc.R
\name{readderivaadc}
\alias{readderivaadc}
\title{Read data from polar climatological and other summary environmental data}
\usage{
readderivaadc(products, xylim = NULL, returnfiles = FALSE, ...)
}
\arguments{
\item{products}{choice of products, see \code{\link{derivaadcproducts}} for available products}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}}

\item{returnfiles}{ignore options and just return the file names}

\item{...}{passed to brick, primarily for \code{filename}}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
Derived data are read from files managed by \code{\link{derivaadcfiles}}.
}
\examples{
\dontrun{
prods <- c("bathymetry","chl_summer_climatology")
x <- readderivaadc(prods)
}
}
\seealso{
\code{\link{derivaadcfiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}

