% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readfastice}
\alias{readfastice}
\title{Fast ice data}
\usage{
readfastice(date, time.resolution = "weekly3", xylim = NULL,
  returnfiles = FALSE, ...)
}
\arguments{
\item{date}{date or dates to read (can be character, POSIXt, or Date)}

\item{time.resolution}{fixed at roughly "3 weekly"}

\item{xylim}{extent in native space of the grid}

\item{returnfiles}{return the file details only}

\item{...}{reserved for future use, currently ignored}
}
\value{
RasterBrick with 1 for fast ice pixels, 0 for other, NA for land mask
}
\description{
Read fast ice data, optionally with a mask
}
\details{
Fast ice data on original Equal Area Cylindrical grid
}
\examples{
r <- readfastice(c("2002-02-10", "2002-03-03"))
}
\references{
\url{http://data.aad.gov.au/aadc/metadata/metadata.cfm?entry_id=modis_20day_fast_ice}
}

