% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsle.R
\name{readfsle}
\alias{readfsle}
\title{Backward-in-time Finite-Size Lyapunov Exponents}
\usage{
readfsle(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  latest = TRUE,
  varname = c("fsle_max", "theta_max"),
  returnfiles = FALSE,
  verbose = TRUE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resolution to read (only daily)}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{latest}{if TRUE and input date is missing return the latest time available, otherwise the earliest}

\item{varname}{either `fsle_max` or `theta_max``}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{verbose}{print messages on progress etc.}

\item{...}{passed to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\value{
data.frame
}
\description{
Daily files
}
