% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrsst.R
\name{readghrsst}
\alias{readghrsst}
\title{Read GHRSST}
\usage{
readghrsst(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  lon180 = TRUE,
  varname = c("analysed_sst", "analysis_error", "mask", "sea_ice_fraction"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{datetime}

\item{time.resolution}{daily}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[terra]{ext}}, see Details}

\item{lon180}{currently ignored}

\item{varname}{one of "analysed_sst", "analysis_error", "mask", "sea_ice_fraction"}

\item{setNA}{ignored}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{arguments passed to raster brick}

\item{inputfiles}{input the files data base to speed up initialization}
}
\value{
SpatRaster, or a data.frame of files if \code{returnfiles = TRUE}
}
\description{
SST in Kelvin
}
