% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrsst.R
\name{readghrsst}
\alias{readghrsst}
\title{Read GHRSST}
\usage{
readghrsst(date, time.resolution = c("daily"), xylim = NULL,
  lon180 = TRUE, varname = c("analysed_sst", "analysis_error", "mask",
  "sea_ice_fraction"), setNA = TRUE, latest = FALSE, returnfiles = FALSE,
  ...)
}
\arguments{
\item{date}{datetime}

\item{time.resolution}{daily}

\item{varname}{one of "analysed_sst", "analysis_error", "mask", "sea_ice_fraction"}

\item{setNA}{ignored}

\item{latest}{return latest available data}

\item{returnfiles}{return the files}

\item{...}{pass in filename to writeRaster}
}
\value{
RasterStack or RasterLayer
}
\description{
Read GHRSST
}

