% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{readice}
\alias{readice}
\alias{readice_daily}
\alias{readice_monthly}
\title{Read from NSIDC 25km polar sea ice.}
\usage{
readice_daily(
  date,
  time.resolution = "daily",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

readice(
  date,
  time.resolution = "daily",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

readice_monthly(
  date,
  time.resolution = "monthly",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{product}{choice of sea ice product, see Details}

\item{hemisphere}{north or south}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}}

\item{setNA}{mask zero and values greater than 100 as NA}

\item{rescale}{rescale values from integer range?}

\item{latest}{if TRUE and date input is missing, return the latest time available otherwise the earliest}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}

\item{extension}{default for product "amsr" is "hdf" but can be "tif" , extension = "hdf"}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
Sea ice at 25km for either hemisphere.
}
\details{
This function relies on the file-listing of \code{\link[=icfiles]{icfiles()}}.

Currently available products are

\describe{
\item{'nsidc'}{daily or monthly NSIDC concentration data, processed by the SMMR/SSMI NASA Team}
}

Dates are matched to file names by finding the nearest match in
time within a short duration. If \code{date} is greater than
length 1 then the sorted set of unique matches is returned.

For NSIDC data a \code{\link[raster]{ratify}}ied raster is returned if \code{setNA} and
\code{rescale} are both set to \code{FALSE}.  Use \code{levels(x)} to return the data.frame of values
and levels (there's no straight-through rule, all numeric values are explicit along with special
values like "Unused").
The values used are documented here \url{http://nsidc.org/data/docs/daac/nsidc0051_gsfc_seaice.gd.html}
}
\examples{
library(raadtools)
 
ice <- readice(latest = TRUE)
}
\seealso{
\code{\link{icefiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}
