% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{readice_area}
\alias{readice_area}
\title{Area of pixels in sea ice}
\usage{
readice_area(product = "nsidc", hemisphere = "south", ...)
}
\arguments{
\item{product}{"nsidc" the 25km NSIDC passive microwave}

\item{hemisphere}{south (default) or north}

\item{...}{ignored}
}
\value{
raster with the area of the cells in m^2
}
\description{
Read the NSIDC pixel-area files for either hemisphere.
Only 25km product is supported. Tool name "psn25area_v3.dat and pss25area_v3.dat":
\url{http://nsidc.org/data/polar-stereo/tools_geo_pixel.html#pixel_area}.
}
\examples{
readice_area()
readice_area(hemisphere = "north")
}
\seealso{
readice
}
