% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid_response.R
\name{readrapid_response}
\alias{readrapid_response}
\title{MODIS Rapid Response images}
\usage{
readrapid_response(date, product = c("aqua", "terra"), latest = FALSE,
  returnfiles = FALSE, ...)
}
\arguments{
\item{date}{date of image to load}

\item{latest}{if TRUE return the latest time available, ignoring the 'date' argument}

\item{returnfiles}{return just the list of files}

\item{...}{other arguments for \code{\link[raster]{brick}}}
}
\description{
Read MODIS Rapid Response RGB images
}
\details{
MODIS Rapid Response, Antarctica
}
