% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity.R
\name{readsal}
\alias{readsal}
\title{Read SMAP sea surface salinity data}
\usage{
readsal(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  lon180 = TRUE,
  varname = c("sss_smap", "nobs", "sss_ref", "gland", "gice", "surtep"),
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read,}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180]}

\item{varname}{variable to return from the data files, default is
"sss_smap", also available is "nobs", "sss_ref", "gland", "gice", "surtep"}

\item{setNA}{mask out land values (only applies to monthly time.resolution)}

\item{latest}{if TRUE (and date not supplied) return the latest time available}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{
  Arguments passed on to \code{\link[=raadtools]{raadtools}}
  \describe{
    \item{\code{}}{}
  }}

\item{inputfiles}{input the files data base to speed up initialization}

\item{readall}{FALSE by default}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
SMAP surface salinity data read from files managed by
\code{\link{salfiles}}.
}
