% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-ssh.R, R/ssh.R
\name{readssh}
\alias{readssh}
\title{Read sea surface height data}
\usage{
readssh(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  lon180 = TRUE,
  ssha = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  verbose = TRUE,
  ...,
  inputfiles = NULL
)

readssh(
  date,
  time.resolution = c("daily"),
  xylim = NULL,
  lon180 = TRUE,
  ssha = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  verbose = TRUE,
  ...,
  inputfiles = NULL
)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resolution to read}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view \link{0, 360} data to Atlantic view \link{-180, 180}
components, in degrees (0 north, 90 east, 180 south, 270 west)}

\item{ssha}{logical, to optionally return anomaly or height}

\item{latest}{if TRUE and date input is missing return the latest time available, otherwise the earliest}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{verbose}{print messages on progress etc.}

\item{...}{passed to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}
}
\value{
\code{RasterBrick} or \code{RasterLayer}, or tibble if \code{returnfiles = TRUE}

data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{readssh} is superseded by \code{\link{read_cmems_adt_daily}} and
\code{\link{read_cmems_sla_daily}}, which return terra \code{SpatRaster} objects.

Sea surface height/anomaly
}
\details{
Details
}
\seealso{
\code{\link{read_cmems_adt_daily}}, \code{\link{read_cmems_sla_daily}}
}
