% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{readsst}
\alias{readsst}
\title{Read OISST sea surface temperature data}
\usage{
readsst(date, time.resolution = c("daily", "monthly"), xylim = NULL,
  lon180 = TRUE, varname = c("sst", "anom", "err", "ice"), setNA = TRUE,
  latest = FALSE, returnfiles = FALSE, ..., inputfiles = NULL)
}
\arguments{
\item{date}{date or dates of data to read, see Details}

\item{time.resolution}{time resoution data to read, daily or monthly}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{lon180}{defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180]}

\item{varname}{variable to return from the data files, default is
"sst" or "anom", "err", "ice"}

\item{setNA}{mask out land values (only applies to monthly time.resolution)}

\item{latest}{if TRUE return the latest time available, ignoring the 'date' argument}

\item{returnfiles}{ignore options and just return the file names and dates}

\item{...}{passed in to brick, primarily for \code{filename}}

\item{inputfiles}{input the files data base to speed up initialization}

\item{readall}{FALSE by default}
}
\value{
\code{\link[raster]{raster}} object
}
\description{
SST data read from files managed by
\code{\link{sstfiles}}. Dates are matched to file names by finding
the nearest match in time within a short duration. If \code{date}
is greater than length 1 then the sorted set of unique matches is
returned.
}
\examples{
\dontrun{
## read one time slice and plot it up in preparation for reading a time series
d <- readsst()
plot(d)
## this step is interactive, draw a boundary on the plot
ext <- drawExtent()
## these can be created manually with xmin,xmax,ymin,ymax
## ext <- extent(-100, 150, -75, -30)
## now read a big chunk of data for this small region
dts <- seq(as.Date("2001-01-03"), by = "1 week", length = 100)
sst <- readsst(dts, xylim = ext)
}

}
\seealso{
\code{\link{icefiles}} for details on the repository of
data files, \code{\link[raster]{raster}} for the return value
}
