% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{readtopo}
\alias{readtopo}
\alias{topofile}
\alias{readbathy}
\alias{readbathy}
\alias{topofile}
\title{Topography data}
\usage{
readtopo(topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
  "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m",
  "macrie2100m"), polar = FALSE, lon180 = TRUE, xylim = NULL,
  returnfiles = FALSE, ...)

readbathy(topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
  "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m",
  "macrie2100m"), polar = FALSE, lon180 = TRUE, xylim = NULL,
  returnfiles = FALSE, ...)

topofile(topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
  "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m",
  "macrie2100m"), polar = FALSE, lon180 = TRUE, ...)
}
\arguments{
\item{topo}{Data source, see Details.}

\item{polar}{Flag for returning the polar version of the IBCSO data.}

\item{lon180}{Flag for returning data in Atlantic [-180, 180] rather than Pacific [0, 360] view.}

\item{xylim}{spatial extents to crop from source data, can be anything accepted by \code{\link[raster]{extent}}, see Details}

\item{returnfiles}{Return just the relevant file name}

\item{...}{reserved for future use, ignored currently}
}
\value{
\describe{
\item{}{\code{topofile} returns a character string of the full path to a file name}
\item{}{\code{readtopo} and \code{readbathy} return the requested data as a RasterLayer (these are aliases)}
}
}
\description{
Functions to provide topographic (bathymetry and/or topography) data.
}
\details{
Use \code{readtopo} (or its alias \code{readbathy}) to read data
from the chosen data set. The function \code{topofile} is used to
find the full file name.
\code{xylim} is expected to be consistent with \code{lon180}
The following data sets are available using the argument \code{topo}.
\describe{
\item{gebco_08}{The GEBCO_08 Grid, a global 30 arc-second grid largely generated by combining quality-controlled ship depth soundings with interpolation between sounding points guided by satellite-derived gravity data. \url{http://www.gebco.net/data_and_products/gridded_bathymetry_data/}}
\item{ibcso}{IBCSO bathymetry data, resolution 1min, use argument \code{polar = TRUE} to return the projected version (polar stereographic with true scale at 71S, WGS84), 500m resolution. \url{http://www.ibcso.org/data.html}}
\item{etopo1}{ETOPO1 is a 1 arc-minute global relief model of Earth's surface that integrates land topography and ocean bathymetry. \url{http://www.ngdc.noaa.gov/mgg/global/global.html}}
\item{etopo2}{Historic and deprecated prior version of ETOPO1. \url{http://www.ngdc.noaa.gov/mgg/global/etopo2.html}}
\item{kerguelen}{Kerguelen Plateau Bathymetric Grid, GeoScience Australia}
\item{george_v_terre_adelie}{A bathymetric Digital Elevation Model (DEM) of the George V and Terre Adelie continental shelf and margin - 100, 250, and 500 metre resolution. \url{http://data.aad.gov.au/aadc/metadata/metadata_redirect.cfm?md=AMD/AU/GVdem_2008}}
\item{smith_sandwell}{Global seafloor topography from satellite altimetry and ship depth soundings. \url{http://topex.ucsd.edu/WWW_html/mar_topo.html}}
}
}
