% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sose.R
\name{sose_monthly_varnames}
\alias{sose_monthly_varnames}
\title{SOSE varnames}
\usage{
sose_monthly_varnames(iteration = "")
}
\value{
chacter vector
}
\description{
Available variable names from SOSE.
}
\details{
These varnames can be used in \code{read_sose(varname = )}.
}
\examples{
sose_monthly_varnames()
}
