% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{sshfiles}
\alias{sshfiles}
\title{AVISO sea surface height / anomaly files}
\usage{
sshfiles(time.resolution = c("daily"), ...)
}
\arguments{
\item{time.resolution}{set to daily only}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of AVISO SSH/SSHA data
}
\details{
A data.frame of file names and dates
}
\seealso{
\code{\link{readssh}}
}
