% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sstfiles}
\alias{sstfiles}
\title{OISST sea surface temperature files}
\usage{
sstfiles(time.resolution = c("daily"), ...)
}
\arguments{
\item{time.resolution}{time resolution read}

\item{...}{reserved for future use, currently ignored}
}
\value{
data.frame of file names and dates
}
\description{
Load file names and dates of OISST sea surface temperature data
}
\details{
A data frame of file names and datres
}

